refactor: split test by types
